requirements.txt
Django==4.0.4
mysql-connector-python==8.0.29
mysqlclient==2.1.0
pantab==2.1.1 # Python bridge between Pandas and Hyper
celery==5.2.6 # Scheduled Jobs
django-model-utils==4.2.0 # Timestamped Models
django-cors-headers==3.12.0
pyhumps==3.7.1 # camelize response keys
SQLAlchemy==1.4.36 # SQL generate
graphene-django==2.15.0 # Graphql
honcho==1.1.0 # Foreman python


# DEV
black==22.3.0 # code formatter
